third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristic/characteristic-found.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['generic_access']}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('generic_access'))
    .then(service => {
      return Promise.all([
          service.getCharacteristic(device_name.alias),
          service.getCharacteristic(device_name.name),
          service.getCharacteristic(device_name.uuid)])
        .then(characteristics => {
          characteristics.forEach(characteristic => {
            assert_equals(
              characteristic.uuid, device_name.uuid,
              'Characteristic UUID should be the same as requested UUID.');
            assert_equals(
              characteristic.service, service,
              'Characteristic service should be the same as service.');
          });
       });
    });
}, 'Request for characteristic. Should return right characteristic');
</script>
